pyproject.toml
[tool.poetry]
name = "mcp-mothership"
version = "0.1.0"
description = "MCP server management platform — orchestrates multiple MCP servers with shared configuration, process management, and logging"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
mcp = ">=1.26.0"
google-genai = ">=1.0.0"
pydantic-settings = "*"
pyyaml = "*"
fastapi = "*"
uvicorn = {version = "*", extras = ["standard"]}

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:ament_lint -p no:ament_flake8 -p no:ament_xmllint -p no:ament_pep257 -p no:ament_copyright -p no:launch_testing -p no:launch_testing_ros"
